Seed annotations for all users
When importing data with annotations (to seed the annotations), doccano either associates
those annotations with the user who uploaded the data (so annotators cannot see it) or
by selecting "Share annotations across users", all users can see it, but cannot delete the seeded
annotations. This change allows uploaded annotations to be made visible to all project members
but still allows them to independently add/remove to their copy of annotations.